feat: refactor styles in tailwind
