+++ src/components/Textbook/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { useState, useEffect } from 'react';
-import Styles from './textbook.module.scss';
 import PageTitle from './PageTitle';
 import SubTitle from './SubTitle';
 import WordDetail from './WordDetail';
@@ -8,7 +7,6 @@ import Empty from './Empty';
 import PageNumber from './PageNumber';
 
 import dictionaryData from './dictionaryData.json';
-import Button from '../Shared/Buttons';
 const Textbook = () => {
   interface Word {
     name: string;
